requirements.txt
torch>=2.1.0
torchvision>=0.16.0
diffusers>=0.29.0
transformers>=4.44.0
accelerate>=0.33.0
safetensors>=0.4.3
peft>=0.17.0
Pillow>=10.0.0
tqdm>=4.66.0
